util.js added and all files updated
