+++ lib/main.js
@@ -20,7 +20,7 @@ exports.zfs = {
     clone: zfs.clone
 };
 
-//ZPOOL EXPORTS
+//ZPooL EXPORTS
 exports.zpool = {
     set: zpool.set,
     destroy: zpool.destroy,
