+++ lib/zfs.js
@@ -1,37 +1,8 @@
 var cp = require('child_process');
-var fs = require('fs');
-var path = require('path');
-var util = require('util');
+var util = require('./util');
 
-if (!fs.existsSync) {
-    fs.existsSync = path.existsSync;
-}
 
-function compact(array) {
-    return array.filter(function (i) {
-        if (typeof i.length !== 'undefined')
-            return i.length > 0;
-        return !!i;
-    });
-}
-
-function findCmd(name) {
-    "use strict";
-
-    var paths = process.env['PATH'].split(':');
-    var pathLen = paths.length;
-    for (var i = 0; i < pathLen; i++) {
-        var sp = path.resolve(paths[i]);
-        var fname = path.normalize(path.join(sp, name));
-        if (fs.existsSync(fname)) {
-            return fname;
-        }
-    }
-
-    return null;
-}
-
-var zfsBin = findCmd('zfs');
+var zfsBin = util.findCmd('zfs');
 function zfs(args, callback) {
     "use strict";
 
@@ -47,28 +18,6 @@ function zfs(args, callback) {
     });
 }
 
-function parseNumber(str) {
-    var m = str.match(/^([0-9.]+)([KMGT]?)$/);
-    if (!m) {
-        return -1;
-    }
-
-    var n = parseFloat(m[1]);
-    if (m[2] === 'K') {
-        n *= 1024;
-    } else if (m[2] === 'M') {
-        n *= 1024 * 1024;
-    } else if (m[2] === 'G') {
-        n *= 1024 * 1024 * 1024;
-    } else if (m[2] === 'T') {
-        n *= 1024 * 1024 * 1024 * 1024;
-    } else if (m[2] === 'P') {
-        n *= 1024 * 1024 * 1024 * 1024 * 1024;
-    }
-
-    return Math.floor(n);
-}
-
 function ZFS(info) {
     "use strict";
 
@@ -82,11 +31,11 @@ function ZFS(info) {
     }
 
     obj.name = info[0];
-    obj.used = parseNumber(info[1]);
+    obj.used = util.parseNumber(info[1]);
     if (info[2] !== '-') { // Snapshots don't have 'avail' field.
-        obj.avail = parseNumber(info[2]);
+        obj.avail = util.parseNumber(info[2]);
     }
-    obj.refer = parseNumber(info[3]);
+    obj.refer = util.parseNumber(info[3]);
     obj.mountpoint = info[4];
 
     Object.freeze(obj);
@@ -140,7 +89,7 @@ function list(opts, cb) {
                 cb(err);
                 return;
             }
-            var lines = compact(stdout.split('\n'));
+            var lines = util.compact(stdout.split('\n'));
             var list = lines.map(function (x) { return new ZFS(x); });
             cb(err, list);
         }
@@ -164,8 +113,8 @@ function get(opts, cb) {
     zfs(params, function (err, stdout) {
         if (cb && typeof cb === 'function') {
             if (err) return cb(err);
-            var lines = compact(stdout.split('\n'));
-            var list = lines.map(function (x) { return new Property(x); });
+            var lines = util.compact(stdout.split('\n'));
+            var list = lines.map(function (x) { return new util.Property(x); });
             cb(err, list);
         }
     });
@@ -214,7 +163,7 @@ function create(opts, cb) {
     }
 
     if (opts.size) {
-      params.push('-V', parseNumber(opts.size));
+      params.push('-V', util.parseNumber(opts.size));
     }
     params.push(opts.name);
