+++ lib/zpool.js
@@ -1,80 +1,9 @@
 var cp = require('child_process');
-var fs = require('fs');
-var path = require('path');
-var util = require('util');
-if (!fs.existsSync) {
-    fs.existsSync = path.existsSync;
-}
-
-function compact(array) {
-    return array.filter(function (i) {
-        if (typeof i.length !== 'undefined')
-            return i.length > 0;
-        return !!i;
-    });
-}
-
-function findCmd(name) {
-    "use strict";
-
-    var paths = process.env['PATH'].split(':');
-    var pathLen = paths.length;
-    for (var i = 0; i < pathLen; i++) {
-        var sp = path.resolve(paths[i]);
-        var fname = path.normalize(path.join(sp, name));
-        if (fs.existsSync(fname)) {
-            return fname;
-        }
-    }
-
-    return null;
-}
+var util = require('./util');
 
 
-function parseNumber(str) {
-    var m = str.match(/^([0-9.]+)([KMGT]?)$/);
-    if (!m) {
-        return -1;
-    }
 
-    var n = parseFloat(m[1]);
-    if (m[2] === 'K') {
-        n *= 1024;
-    } else if (m[2] === 'M') {
-        n *= 1024 * 1024;
-    } else if (m[2] === 'G') {
-        n *= 1024 * 1024 * 1024;
-    } else if (m[2] === 'T') {
-        n *= 1024 * 1024 * 1024 * 1024;
-    } else if (m[2] === 'P') {
-        n *= 1024 * 1024 * 1024 * 1024 * 1024;
-    }
-
-    return Math.floor(n);
-}
-
-function Property(info) {
-    "use strict";
-
-    var obj = Object.create({});
-    if (typeof info === 'string') {
-        info = info.split(/\s+/, 4);
-    }
-
-    if (info.length !== 4) {
-        return null;
-    }
-
-    obj.name = info[0];
-    obj.property = info[1];
-    obj.value = info[2];
-    obj.source = info[3];
-
-    Object.freeze(obj);
-    return obj;
-}
-
-function ZPOOL(info) {
+function ZPool(info) {
     "use strict";
 
     var obj = Object.create({});
@@ -88,11 +17,8 @@ function ZPOOL(info) {
 
     obj.name = info[0];
     obj.size = info[1];
-    obj.sizeAbsolute = parseNumber(info[1]);
     obj.alloc = info[2];
-    obj.sizeAbsolute = parseNumber(info[2]);
     obj.free = info[3];
-    obj.freeAbsolute = parseNumber(info[3]);
     obj.cap = info[4];
     obj.health = info[5];
     if (info[6] !== '-') {
@@ -120,14 +46,14 @@ function analyzeVdevOpts (deviceList) {
     return devs;
 }
 
-var zpoolBin = findCmd('zpool');
+var zpoolBin = util.findCmd('zpool');
 function zpool(args, callback) {
     "use strict";
 
     cp.execFile(zpoolBin, args, {maxBuffer: 8000000}, function (err, stdout, stderr) {
         if (callback && typeof callback === 'function') {
             if (err) {
-                err.message = compact(err.message.split('\n')).join('; ').trim();
+                err.message = util.compact(err.message.split('\n')).join('; ').trim();
                 callback(err);
             } else {
                 if (!stdout || stdout == '') stdout = 'Request succesfully fulfilled!';
@@ -238,8 +164,8 @@ function list(opts, cb) {
                 cb(err);
                 return;
             }
-            var lines = compact(stdout.split('\n'));
-            var list = lines.map(function (x) { return new ZPOOL(x); });
+            var lines = util.compact(stdout.split('\n'));
+            var list = lines.map(function (x) { return new ZPool(x); });
             cb(err, list);
         }
     });
@@ -263,8 +189,8 @@ function get(opts, cb) {
     zpool(params, function (err, stdout) {
         if (cb && typeof cb === 'function') {
             if (err) return cb(err);
-            var lines = compact(stdout.split('\n'));
-            var list = lines.map(function (x) { return new Property(x); });
+            var lines = util.compact(stdout.split('\n'));
+            var list = lines.map(function (x) { return new util.Property(x); });
             cb(err, list);
         }
     });
